[REFACTOR] : 음성기능 커스텀 훅 useRecording 변경 및 ui/ux 개선
