+++ src/pages/Chat/index.tsx
@@ -2,13 +2,13 @@ import React, { useState, useRef, useEffect } from 'react';
 import { TbSend } from 'react-icons/tb';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { BsFillMicFill } from 'react-icons/bs';
 import Button from '@/components/Common/Button';
 import ChatLoading from '@/components/Chat/ChatLoading';
 import MarkDown from '@/components/Chat/MarkDown';
 import { API_URLS, BASE_URL, ROUTE_PATHS } from '@/constants/config';
 import Header from '@/components/Common/Header';
 import ImageUploadButton from '@/components/Chat/ImageUploadButton';
-import SpeechToTextButton from '@/components/Chat/SpeechToTextButton';
 import gildong from '@/assets/gildong_3d.png';
 import AddItineraryButton from '@/components/Travel/AddItineraryButton';
 import { imageState } from '@/store/atom/travelAtom';
@@ -22,6 +22,8 @@ import beach from '@/assets/beach.png';
 import blindPerson from '@/assets/blind_person.png';
 import mountain from '@/assets/mountain.png';
 import wheelchair from '@/assets/wheelchair.png';
+import voiceLoading from '@/assets/loading_voice.gif';
+import useRecording from '@/hooks/useRecording';
 import styles from './styles.module.scss';
 interface ChatProps {
   home?: boolean;
@@ -41,10 +43,15 @@ export default function Chat({ home }: ChatProps) {
   const [isChatLoading, setIsChatLoading] = useState(false);
   const [question, setQuestion] = useState('');
   const [answer, setAnswer] = useState('');
-  const [isMicOn, setIsMicOn] = useState(false);
-  const [isMicLoading, setIsMicLoading] = useState(false);
   const [isImageOpen, setIsImageOpen] = useState(false);
   const setPage = useSetRecoilState(pageState);
+  const {
+    isRecording,
+    isSTTLoading,
+    isRecordingStarting,
+    startRecording,
+    stopRecording,
+  } = useRecording(setValue);
 
   const handleInput = (event: React.ChangeEvent<HTMLInputElement>) => {
     setValue(event.target.value);
@@ -330,28 +337,32 @@ export default function Chat({ home }: ChatProps) {
                       </div>
                     </div>
                   ) : null}
+                  {isRecording ? (
+                    <div className={styles.voiceContainer}>
+                      <div className={styles.voice}>
+                        듣고 있습니다.
+                        <div className={styles.pulse} onClick={stopRecording}>
+                          <div className={styles.rectangle} />
+                        </div>
+                      </div>
+                    </div>
+                  ) : isSTTLoading ? (
+                    <div className={styles.voiceContainer}>
+                      <div className={styles.voice}>
+                        <img src={voiceLoading} />
+                      </div>
+                    </div>
+                  ) : null}
                   <div className={styles.icon}>
                     <ImageUploadButton setIsImageOpen={setIsImageOpen} />
                   </div>
-                  {isMicOn ? (
-                    <div className={styles.textWrapper}>
-                      <div className={styles.typingText}>
-                        듣고 있습니다. 마이크에 대고 계속 말해주세요!
-                      </div>
-                    </div>
-                  ) : isMicLoading ? (
-                    <div className={styles.textWrapper}>
-                      <ChatLoading />
-                    </div>
-                  ) : (
-                    <input
-                      className={styles.input}
-                      onChange={handleInput}
-                      value={value}
-                      onKeyDown={handleEnter}
-                      placeholder="무엇이든 물어보세요!"
-                    />
-                  )}
+                  <input
+                    className={styles.input}
+                    onChange={handleInput}
+                    value={value}
+                    onKeyDown={handleEnter}
+                    placeholder="무엇이든 물어보세요!"
+                  />
                   <div className={styles.send}>
                     {value || isImageOpen ? (
                       <Button
@@ -362,14 +373,17 @@ export default function Chat({ home }: ChatProps) {
                         variant="primary"
                         onClick={handleSubmit}
                       />
-                    ) : (
-                      <SpeechToTextButton
-                        setValue={setValue}
-                        isMicOn={isMicOn}
-                        setIsMicOn={setIsMicOn}
-                        setIsMicLoading={setIsMicLoading}
-                      />
-                    )}
+                    ) : !isRecordingStarting ? (
+                      <Button
+                        icon={<BsFillMicFill />}
+                        size="sm"
+                        color="secondary"
+                        iconBtn={true}
+                        onClick={startRecording}
+                      >
+                        녹음 시작
+                      </Button>
+                    ) : null}
                   </div>
                 </div>
               </div>
